requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
fastapi[all]
uvicorn[standard]
gunicorn

asyncpg

sqlalchemy
sqlalchemy

pydantic
starlette
strawberry-graphql
click
aiodataloader

https://github.com/hrbolek/uoishelpers/archive/refs/heads/main.zip
